delete all binaries on dsf fhir server if transfer error occurs

+++ src/main/java/de/medizininformatik_initiative/process/data_transfer/service/DeleteData.java
@@ -6,6 +6,8 @@ import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.hl7.fhir.r4.model.Binary;
 import org.hl7.fhir.r4.model.DocumentReference;
 import org.hl7.fhir.r4.model.IdType;
+import org.hl7.fhir.r4.model.ListResource;
+import org.hl7.fhir.r4.model.Reference;
 import org.hl7.fhir.r4.model.Resource;
 import org.hl7.fhir.r4.model.Task;
 import org.slf4j.Logger;
@@ -35,37 +37,39 @@ public class DeleteData extends AbstractServiceDelegate
 		String projectIdentifier = variables
 				.getString(ConstantsDataTransfer.BPMN_EXECUTION_VARIABLE_PROJECT_IDENTIFIER);
 
-		DocumentReference documentReference = variables
-				.getResource(ConstantsDataTransfer.BPMN_EXECUTION_VARIABLE_TRANSFER_DOCUMENT_REFERENCE);
+		String transferDocumentReferenceLocation = variables
+				.getString(ConstantsDataTransfer.BPMN_EXECUTION_VARIABLE_TRANSFER_DOCUMENT_REFERENCE_LOCATION);
+		ListResource transferBinaryReferenceList = variables
+				.getResource(ConstantsDataTransfer.BPMN_EXECUTION_VARIABLE_TRANSFER_DATA_RESOURCES);
 
 		logger.info(
 				"Permanently deleting data-set provided for DMS '{}' and project-identifier '{}' referenced in Task with id '{}' (DocumentReference with id '{}' and its encrypted attachments)",
-				dmsIdentifier, projectIdentifier, task.getId(), documentReference.getId());
+				dmsIdentifier, projectIdentifier, task.getId(), transferDocumentReferenceLocation);
 
 		try
 		{
-			List<IdType> attachments = getAttachmentIds(documentReference);
+			List<IdType> attachments = getAttachmentIds(transferBinaryReferenceList);
 
 			deletePermanently(attachments, Binary.class);
-			deletePermanently(documentReference.getIdElement(), DocumentReference.class);
+			deletePermanently(new IdType(transferDocumentReferenceLocation), DocumentReference.class);
 		}
 		catch (Exception exception)
 		{
 			logger.warn(
 					"Could not permanently delete data-set for DMS '{}' and project-identifier '{}' referenced in Task with id '{}' (DocumentReference with id '{}' and its encrypted attachments) - {}",
-					dmsIdentifier, projectIdentifier, task.getId(), documentReference.getId(), exception.getMessage());
+					dmsIdentifier, projectIdentifier, task.getId(), transferDocumentReferenceLocation,
+					exception.getMessage());
 
 			String error = "Permanently deleting encrypted data-set failed - " + exception.getMessage();
 			throw new RuntimeException(error, exception);
 		}
 	}
 
-	private List<IdType> getAttachmentIds(DocumentReference documentReference)
+	private List<IdType> getAttachmentIds(ListResource transferBinaryReferenceList)
 	{
-		return documentReference.getContent().stream()
-				.filter(DocumentReference.DocumentReferenceContentComponent::hasAttachment)
-				.map(DocumentReference.DocumentReferenceContentComponent::getAttachment)
-				.map(a -> new IdType(a.getUrl())).toList();
+		return transferBinaryReferenceList.getEntry().stream().filter(ListResource.ListEntryComponent::hasItem)
+				.map(ListResource.ListEntryComponent::getItem).filter(Reference::hasReference)
+				.map(i -> (IdType) i.getReferenceElement()).toList();
 	}
 
 	private void deletePermanently(List<IdType> idTypes, Class<? extends Resource> resourceType)
@@ -75,9 +79,13 @@ public class DeleteData extends AbstractServiceDelegate
 
 	private void deletePermanently(IdType idType, Class<? extends Resource> resourceType)
 	{
-		BasicFhirWebserviceClient client = api.getFhirWebserviceClientProvider().getLocalWebserviceClient()
-				.withRetry(ConstantsBase.DSF_CLIENT_RETRY_6_TIMES, ConstantsBase.DSF_CLIENT_RETRY_INTERVAL_5MIN);
-		client.delete(resourceType, idType.getIdPart());
-		client.deletePermanently(resourceType, idType.getIdPart());
+		idType = idType.toVersionless();
+		if (idType.hasIdPart())
+		{
+			BasicFhirWebserviceClient client = api.getFhirWebserviceClientProvider().getLocalWebserviceClient()
+					.withRetry(ConstantsBase.DSF_CLIENT_RETRY_6_TIMES, ConstantsBase.DSF_CLIENT_RETRY_INTERVAL_5MIN);
+			client.delete(resourceType, idType.getIdPart());
+			client.deletePermanently(resourceType, idType.getIdPart());
+		}
 	}
 }
